requirements-app.txt
absl-py==0.8.1
astor==0.8.0
beautifulsoup4==4.8.1
blis==0.4.1
boto==2.49.0
boto3==1.10.34
botocore==1.13.34
bs4==0.0.1
cachetools==3.1.1
catalogue==0.0.8
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
cymem==2.0.3
Cython==0.29.12
dataclasses==0.7
DAWG-Python==0.7.2
deeppavlov==0.7.1
docopt==0.6.2
docutils==0.15.2
en-core-web-md==2.0.0
en-core-web-sm==2.2.5
fastapi==0.38.1
Flask==1.1.1
fuzzywuzzy==0.17.0
gast==0.2.2
gensim==3.8.1
google-api-core==1.14.3
google-auth==1.7.2
google-auth-oauthlib==0.4.1
google-cloud-core==0.29.1
google-cloud-translate==1.4.0
google-pasta==0.1.8
googleapis-common-protos==1.6.0
grpcio==1.25.0
h11==0.8.1
h5py==2.9.0
html5lib==1.0.1
httptools==0.0.13
idna==2.8
importlib-metadata==1.2.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==8.0.2
murmurhash==1.0.2
nltk>=3.4.5
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
overrides==1.9
pandas==0.24.2
plac==1.1.3
preshed==3.0.2
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pydantic==0.32.2
pygame==1.9.6
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.404381.4453942
pyOpenSSL==19.0.0
pyTelegramBotAPI==3.6.6
python-dateutil==2.8.0
pytorch-transformers==1.0.0
pytz==2019.3
PyYAML==5.2
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
rusenttokenize==0.0.5
s3transfer==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
sentencepiece==0.1.83
six==1.13.0
sklearn==0.0
smart-open==1.9.0
sner==0.2.12
soupsieve==1.9.5
spacy==2.2.3
srsly==0.2.0
starlette==0.12.8
tensorboard==2.0.2
tensorflow==2.0.1
tensorflow-estimator==2.0.1
termcolor==1.1.0
thinc==7.3.1
torch==1.3.1
tqdm==4.32.2
urllib3==1.25.7
uvicorn==0.11.7
uvloop==0.14.0
wasabi==0.4.2
webencodings==0.5.1
websockets==8.1
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
